+++ src/central_node_database.cc
@@ -69,6 +69,7 @@ MpsDb::MpsDb(uint32_t inputUpdateTimeout)
     _pcChangeSameTagCounter(0),
     _pcChangeFirstPacket(true),
     _pcChangeDebug(false),
+    _reloadInactive(false),
     _pcFlagsCounters(Firmware::PcChangePacketFlagsLabels.size(), 0),
     mitigationTxTime( "Mitigation Transmission time", 360 )
 {
@@ -201,13 +202,17 @@ void MpsDb::updateInputs()
         }
 
         _inputUpdateTime.start();
-
+        // If an application card has been set inactive, its logic
+        // will also be set to ignored, so the FW
+        // configuration will need to be reloaded
         DbApplicationCardMap::iterator applicationCardIt;
         for (applicationCardIt = applicationCards->begin();
             applicationCardIt != applicationCards->end();
             ++applicationCardIt)
         {
-            (*applicationCardIt).second->updateInputs();
+            if ((*applicationCardIt).second->updateInputs()){
+                _reloadInactive = true;
+            }
         }
 
         _updateCounter++;